+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
-## v1.0.0
+## v0.9.0
 
-- Initial Release
+[Breaking Change] Refactor `GetSession` method to allow adding the data source config and the user agent to configure the default HTTP client.
+
+## v0.8.0
+
+Added interfaces and functions for SQL data sources